refactor: tweak grouping of views
About to do some permissions-related changes - I think clearer to have them
grouped this way.
