+++ compose.yaml
@@ -53,7 +53,7 @@ services:
        - ./scl/core/templates:/app/scl/core/templates
        - ./scl/core/migrations:/app/scl/core/migrations
        - ./scl/core/static:/app/scl/core/static
-       - ./scl/core/views.py:/app/scl/core/views.py
+       - ./scl/core/views:/app/scl/core/views
        - ./scl/core/models.py:/app/scl/core/models.py
        - ./scl/core/urls.py:/app/scl/core/urls.py
     profiles: ['']
