+++ scl/core/constants.py
@@ -270,3 +270,6 @@ COUNTRIES_AND_TERRITORIES = (
     ("ZM","Zambia"),
     ("ZW","Zimbabwe"),
 )
+
+
+BASIC_ACCESS_GROUP = 'Basic access'
