+++ scl/core/views/healthcheck.py
@@ -1,0 +1,4 @@
+from django.http import HttpResponse, JsonResponse
+
+def lb_healthcheck(request):
+    return HttpResponse("OK")
