+++ scl/core/views/html.py
@@ -5,17 +5,12 @@ import boto3
 from datetime import date
 
 from django.conf import settings
-from django.http import HttpResponse, JsonResponse
 from django.shortcuts import render
 from reversion.models import Version
 
-from .models import Company, Engagement
+from scl.core.models import Company, Engagement
 
 logger = logging.getLogger().warning
-
-
-def lb_healthcheck(request):
-    return HttpResponse("OK")
 
 
 def index(request):
@@ -30,40 +25,6 @@ def index(request):
         "your_companies": your_companies,
         "your_future_enagements": your_future_enagements,
     })
-
-
-def engagement(request):
-    return render(request, "engagement.html")
-
-
-def aws_credentials(request):
-    client = boto3.client("sts")
-    role_arn = settings.AWS_TRANSCRIBE_ROLE_ARN
-
-    # Creating new credentials unfortunately sometimes fails
-    max_attempts = 3
-    for i in range(0, 3):
-        try:
-            credentials = client.assume_role(
-                RoleArn=role_arn,
-                RoleSessionName="scl_" + str(uuid.uuid4()),
-                DurationSeconds=60 * 60,
-            )["Credentials"]
-        except Exception:
-            if i == max_attempts - 1:
-                raise
-            else:
-                time.sleep(1)
-
-    return JsonResponse(
-        {
-            "AccessKeyId": credentials["AccessKeyId"],
-            "SecretAccessKey": credentials["SecretAccessKey"],
-            "SessionToken": credentials["SessionToken"],
-            "Expiration": credentials["Expiration"],
-        },
-        status=200,
-    )
 
 
 def company_briefing(request, duns_number):
