+++ scl/urls.py
@@ -20,20 +20,28 @@ from django.conf import settings
 from django.contrib import admin
 from django.urls import include, path
 
-from scl.core.views import lb_healthcheck, index, aws_credentials
-from scl.core.views import index, company_briefing, engagement
-from scl.core.api.v1.company import company_api
-from scl.core.api.v1.engagement import engagement_api
+from scl.core.views.html import index, company_briefing, engagement
+from scl.core.views.api import aws_credentials_api, company_api, engagement_api
+from scl.core.views.healthcheck import lb_healthcheck
+
 
 urlpatterns = [
-    path('lb-healthcheck', lb_healthcheck),
+    # Auth (SSO)
+    path('auth/', include('authbroker_client.urls')),
+
+    # Admin
     path('admin/', admin.site.urls),
+
+    # HTML
     path("", index, name="home-page"),
-    path("company-briefing/<str:duns_number>",
-         company_briefing, name='company-briefing'),
+    path("company-briefing/<str:duns_number>", company_briefing, name='company-briefing'),
     path("engagement/<uuid:engagement_id>", engagement, name='engagement'),
+
+    # API
     path("api/v1/company/<str:duns_number>", company_api),
     path("api/v1/engagement/<uuid:engagement_id>", engagement_api),
-    path("api/v1/aws-credentials", aws_credentials),
-    path('auth/', include('authbroker_client.urls'))
+    path("api/v1/aws-credentials", aws_credentials_api),
+
+    # Healthcheck
+    path('lb-healthcheck', lb_healthcheck),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
